Add team repository insert method

+++ internal/app/team/type.go
@@ -3,15 +3,15 @@ package team
 import "time"
 
 type Rating struct {
-	TeamID    uint8
-	FixtureID int8
-	SeasonID  int8
+	TeamID    uint64
+	FixtureID uint64
+	SeasonID  uint64
 	Attack    Points
 	Defence   Points
 	Timestamp time.Time
 }
 
 type Points struct {
-	Total      uint64
-	Difference uint64
+	Total      int64
+	Difference int64
 }
